+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 Minor features that don't break backwards compatibility are released as patches.
 
+## 0.5.0
+
+### Added
+
+- Introduce hidden events
+- Trigger "new event tree" event via event pool
+
+### Changed
+
+- Split event trees up by request event
+- Replace low dependency with providers
+
 ## 0.4.0
 
 ### Added